Fixes curriculum manager env reset counter update in each collection round (#58)
# Description

Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context.
List any dependencies that are required for this change.

- Instead of resetting curriculum_manager.env_reset_counter to 0 in each
collection round after resetting all the environments, set it to the
value before the resets.
--**Why?** E.g. if the update_interval is set to 60000, the curriculum
manager will never update the ratios of the mixed agent if each
collection round yields < 60000 resets. For example see images below.

Fixes # (issue)
- Store current number of resets since last ratio update before
resetting all the environments. Then set the
curriculum.env_reset_counter to this value after resetting all envs in
the start of each collection round.

<!-- As a practice, it is recommended to open an issue to have
discussions on the proposed pull request.
This makes it easier for the community to keep track of what is being
developed or added, and if a given feature
is demanded by more than one party. -->

## Type of change

<!-- As you go through the list, delete the ones that are not
applicable. -->

- Bug fix (non-breaking change which fixes an issue)

## Screenshots

Please attach before and after screenshots of the change if applicable.


![image](https://github.com/user-attachments/assets/8c7abfd8-c125-4b9b-87c1-850bc7bddf8e)


<!--
Example:

| Before | After |
| ------ | ----- |
| _gif/png before_ | _gif/png after_ |

To upload images to a PR -- simply drag and drop an image while in edit
mode and it should upload the image directly. You can then paste that
source into the above before/after sections.
-->

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`pre-commit run --all-files`
- [x] My changes generate no new warnings
- [ ] When modifying the omniverse extension, I have updated the
changelog and the corresponding version in the extension's
`config/extension.toml` file

<!--
As you go through the checklist above, you can mark something as done by
putting an x character in it

For example,
- [x] I have done this task
- [ ] I have not done this task
-->