Fixes filtering of small motion trajectories in `TrajectoryDataset` (#72)
# Description

Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context.
List any dependencies that are required for this change.
- Change filtering of small motion trajectories in `populate` call of
`TrajectoryDataset`.
- **Issue**: Assume we have 1000 samples and
`self.cfg.small_motion_ratio=0.1`. Then it doesn't matter if we have 101
small motion trajectories or 500 small motion trajectories. We remove
0.1*1000=100 small motion samples in both cases. This does not produce
the desired ratios as `(101 - 100) / (1000 - 100) = 0.001` and `(500 -
100) / (1000 - 100) = 0.444`.
- **Fix:** Let `x` be the number of small motion samples, `N` the total
number of samples and `r` is the desired ratio. Assume also that `(x /
N) > r` so we need to remove small motion samples. Then let `n` denote
the number of samples to filter out. We want to solve `(x - n) / (N - n)
= r`, which gives `n = (r*N - x) / (r - 1)`.

Fixes # (issue)
- Instead of removing
-- `int(self.states.shape[0] * self.cfg.small_motion_ratio)` small
motion samples.
- Remove
-- `int((self.cfg.small_motion_ratio * self.states.shape[0] -
small_movement_idx.shape[0]) / (self.cfg.small_motion_ratio - 1))` small
motion samples.

<!-- As a practice, it is recommended to open an issue to have
discussions on the proposed pull request.
This makes it easier for the community to keep track of what is being
developed or added, and if a given feature
is demanded by more than one party. -->

## Type of change

<!-- As you go through the list, delete the ones that are not
applicable. -->

- Bug fix (non-breaking change which fixes an issue)

## Screenshots

Please attach before and after screenshots of the change if applicable.

![image](https://github.com/user-attachments/assets/818f0705-fc88-4bf2-aa5a-fcee336fcd80)

![image](https://github.com/user-attachments/assets/4da05cb8-5b53-4bf3-beb9-477df3c72506)

![image](https://github.com/user-attachments/assets/164fec8a-b344-4df5-ba43-1a6d1843e250)

![image](https://github.com/user-attachments/assets/1d68c9d4-431e-40b9-85bb-212e86ea5fa3)

![image](https://github.com/user-attachments/assets/4e795737-079b-4a18-bafe-d16210e6a96f)


![image](https://github.com/user-attachments/assets/c758bdac-b21e-4e66-b0ea-42d84bebfdfd)

![image](https://github.com/user-attachments/assets/d6d29c39-8dde-44d8-93d6-e5cfcd83bcdb)

![image](https://github.com/user-attachments/assets/ccf38c12-aa5b-45f8-8f91-0987d07ada96)

![image](https://github.com/user-attachments/assets/666bb994-25ce-435f-b547-cb645daf02d3)

![image](https://github.com/user-attachments/assets/be06c7aa-200a-4dea-9e75-6ccb1c55ee5e)




<!--
Example:

| Before | After |
| ------ | ----- |
| _gif/png before_ | _gif/png after_ |

To upload images to a PR -- simply drag and drop an image while in edit
mode and it should upload the image directly. You can then paste that
source into the above before/after sections.
-->

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`pre-commit run --all-files`
- [x] My changes generate no new warnings
- [ ] When modifying the omniverse extension, I have updated the
changelog and the corresponding version in the extension's
`config/extension.toml` file

<!--
As you go through the checklist above, you can mark something as done by
putting an x character in it

For example,
- [x] I have done this task
- [ ] I have not done this task
-->